introduce enos_action variable.
This is used to choose which action will be performed during the run of
a role.

For now, only enos_action=deploy is supported and means that the
role(component) must be deployed.

Future action will be supported like backup (backup the state of the
component if relevant) and destroy (destroy the component)