Re-enable qdr for test_case1 (and fun and profit)
* Why:

We want to support qdr *and* rabbitmq as bus backends.
We want to support qdr *and* rabbitmq as control bus backend.
We want to allow failover addresses for ombt agents.
We want to distribute ombt agents on available messaging agent (e.g
routers).

* How:

I generalized the fact that most of the things are generated in python.
Bus configuration is abstrated by the mean of two classes (RabbitMQConf
and QdrConf). This allows the generation of the configuration for the
OmbtAgent. OmbtAgents are also abstracted in various classes
(Ombt[Client|Server|Controller]) allowing their configuration to be also
generated in python.

Those object are saved in the environment when it's necessary. At some
point they are also passed to Ansible, but must be in a dict
representation. In my opinion this add some unnecessary noise to the
code and maybe this should be handled automatically by *run_ansible*.

* What:

* In `prepare` python representation of the configuration of the bus and
control-bus agents are generated : `bus_conf` and  `control_bus_conf`.

* This two objects are saved in the environment and also passed to
Ansible to configure the two buses.

* In `test_case_1` we load these configurations to distribute the
clients amongst all the messaging elements.

* What is missing:

* Failover address isn't implemented yet (normally just need to pass several
element to the ombt agent constructor)
* Intensive testing :)