Add collectd-qdrouterd plugin
To address this integration, I choose to package in the same docker
qdrouterd and collectd together.This probably comes  at the cost of an
overhead for running the qdrouterd containers. Metrics are send to
InfluxDB, the host is set to the id of the router.

Notable change:
- ombt-orchestrator is the name of the influxdb database which gathers
metrics from telegraf and collectd
- docker images generation is added in the repository
- add `qdr_version` in config to choose the docker image to deploy
(with or withour collectd)