Extract method for the bus configuration
This extracts the method that generates the configuration for the
various buses. Initially this was done for enabling the user to choose
its preferred control-bus from the configuration. The goal was to scale
it using qdr.
But I propose to keep the RabbitMQ control bus as the only possible
choice for now as it seems we can scale the control-bus by replicating
it and distributing the agents on one of the shard.