Add support for RabbitMQ cluster
The configuration of the bus with `type:rabbit` now supports:
- a `number` corresponding to the number of rabbitmq instance to
start
- a `mode` whose values can be `standalone` of `cluster`.
Standalone will start `number` independant instances of rmq whereas
`cluster` will linked them to form a cluster.

Notes:
- The control bus is forced to be `standalone` in this
implementation. Nothing prevent to use something else in the future.
The number of instances of the control bus is set to the number of nodes
in the control-bus roles, by default.
- In cluster mode several agents can sits on the same nodes
transparently. It corresponds to the case where `number` is greater
than the number of machines in the group `bus`.

Fix #95