agents<machines led to non-homogeneous distribution
Situation before.

Considering the case where we have 10 machines, 4 shards, 20 servers to start.

The first shard starts 5 servers on machine01-machine05
The second shard starts 5 servers on machine02-machine06
etc
leading to 1-2-3-4-4-3-2-1-0-0 servers being started on the machines.

This was because we only shift the machines using the shard_index whose values
was 0,1,2,3

Situation now.

A shard_index is defined per type of agent (controller, server, client)
For each shard it is set to the total number of agents already distributed.
This will shift the machines more agressively.

Following the previous example:
- the first 5 machines will get 1 server each for the 1st shard (shard_index=0)
- the next 5 machines will get 1 server each for the 2nd shard (shard_index=5)
- the next 5 machines (= the first 5 machines) will get 1 more
server each (shard_index=10)
etc
leading to 2-2-2-2-2-2-2-2-2-2 servers being starting on the machines.